Renamed `SettingsPanel` to `SettingsForm`.
The name now reflects the type of the class, not its supposed usage. The form class itself is oblivious to the fact whether it's displayed as modal "panel" or embedded form and the naming is more consistent.
